Rename app and bundle id.
