+++ KukaApp/android/app/src/main/java/org/codeforhawaii/kuka/MainActivity.java
@@ -1,4 +1,4 @@
-package com.kukaapp;
+package org.codeforhawaii.kuka;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "KukaApp";
+    return "Kuka";
   }
 }
