+++ KukaApp/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "KukaApp",
-  "displayName": "KukaApp"
+  "name": "Kuka",
+  "displayName": "Kuka"
 }
