+++ KukaApp/ios/Kuka-tvOSTests/Info.plist
