+++ KukaApp/ios/Kuka.xcodeproj/xcshareddata/xcschemes/Kuka.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "KukaApp.app"
-               BlueprintName = "KukaApp"
-               ReferencedContainer = "container:KukaApp.xcodeproj">
+               BuildableName = "Kuka.app"
+               BlueprintName = "Kuka"
+               ReferencedContainer = "container:Kuka.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "KukaAppTests.xctest"
-               BlueprintName = "KukaAppTests"
-               ReferencedContainer = "container:KukaApp.xcodeproj">
+               BuildableName = "KukaTests.xctest"
+               BlueprintName = "KukaTests"
+               ReferencedContainer = "container:Kuka.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "KukaApp.app"
-            BlueprintName = "KukaApp"
-            ReferencedContainer = "container:KukaApp.xcodeproj">
+            BuildableName = "Kuka.app"
+            BlueprintName = "Kuka"
+            ReferencedContainer = "container:Kuka.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "KukaApp.app"
-            BlueprintName = "KukaApp"
-            ReferencedContainer = "container:KukaApp.xcodeproj">
+            BuildableName = "Kuka.app"
+            BlueprintName = "Kuka"
+            ReferencedContainer = "container:Kuka.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
