+++ KukaApp/ios/Kuka/AppDelegate.h
