+++ KukaApp/ios/Kuka/Images.xcassets/AppIcon.appiconset/Contents.json
