+++ KukaApp/ios/Kuka/Images.xcassets/Contents.json
