+++ KukaApp/ios/KukaTests/Info.plist
