+++ KukaApp/ios/KukaTests/KukaTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface KukaAppTests : XCTestCase
+@interface KukaTests : XCTestCase
 
 @end
 
-@implementation KukaAppTests
+@implementation KukaTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
