+++ KukaApp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "KukaApp",
+  "name": "Kuka",
   "version": "0.0.1",
   "private": true,
   "scripts": {
